renames config to context
