+++ cmd/harbor/root/context/cmd.go
@@ -11,7 +11,6 @@
 // WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 // See the License for the specific language governing permissions and
 // limitations under the License.
-
 package context
 
 import "github.com/spf13/cobra"
@@ -21,10 +20,15 @@ func Context() *cobra.Command {
 		Use:     "context",
 		Short:   "Manage locally available contexts",
 		Example: "harbor context list",
+		Long: `The context command allows you to manage configs of the Harbor CLI.
+				You can add, get, or delete specific config item, as well as list all config items of the Harbor Cli`,
 	}
 
 	cmd.AddCommand(
 		ListContextCommand(),
+		GetContextItemCommand(),
+		UpdateContextItemCommand(),
+		DeleteContextItemCommand(),
 	)
 
 	return cmd
